SDL-2.0.5: apply fix https://bugzilla.libsdl.org/show_bug.cgi?id=3685 :
Solra Bina		2017-06-27 22:27:48 UTC

With a NULL callback, SDL_OpenAudioDevice sets up the device to be
fed with SDL_QueueAudio. It attempts to allocate at least enough
SDL_AudioBufferQueues to store one buffer of audio data.
Unfortunately, due to a small logic error, if the buffer size is
not an exact multiple of SDL_AUDIOBUFFERQUEUE_PACKETLEN,
SDL_AUDIOBUFFERQUEUE_PACKETLEN additional queues will be allocated
instead of the intended 1 additional queue. In my program, this
resulted in the allocation of 67MB worth of unnecessary queues,
and was singlehandedly responsible for 90% of the program's total
memory usage.

With the attached patch, the extra memory usage is gone and my
program still works. A few cursory attempts to exercise any broken
code paths that may have been unearthed by the patch bore no fruit.

Severity is minor because this bug does not cause crashes, security
risks, glitches, or broken functionality. The only visible effect
is the extra memory usage, which would normally only be a problem
on embedded platforms and Emscripten.